重构: 将 newtrace 移动到 libytht 中
