+++ include/ytht/msg.h
@@ -1,0 +1,11 @@
+#ifndef YTHT_MSG_H
+#define YTHT_MSG_H
+
+struct mymsgbuf {
+	long int mtype;
+	char *mtext;
+};
+
+void newtrace(const char *s);
+#endif
+
